+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 From version 1.2.0 the file IP2LOCATION-LITE-DB5.BIN is no longer part of the docker image and will need to be downloaded (after registration) from [ip2location](https://lite.ip2location.com/database/ip-country?lang=en_US) and be placed in /srv/incore_auth.
 
+# [1.7.0] - 2023-06-14
+
+## Added
+- New user default usage to zero [#38](https://github.com/IN-CORE/incore-auth/issues/38)
+- Return a user object that contains username, fullname, email, groups and roles.
+
+
 # [1.6.0] - 2023-03-14
 
 ## Added